[REFACTOR] service 로직 분리
* 알림 저장 및 조회 로직
* SSE 연결 및 해제 로직
